Moving the openssl based generate_broker_key logic into the base AuthService class
While authentication plugin can override these methods we should not encourage
it as it has the potential to break node->broker authentication when swapping
plugins.
