+++ stickshift/controller/lib/stickshift-controller/config/initializers/broker_key.rb
@@ -1,0 +1,8 @@
+Broker::Application.configure do
+  config.auth = {
+    :salt           => "ClWqe5zKtEW4CJEMyjzQ",
+    :privkeyfile    => "/etc/stickshift/server_priv.pem",
+    :privkeypass    => "",
+    :pubkeyfile     => "/etc/stickshift/server_pub.pem",
+  }
+end
