+++ swingshift/remote-user/lib/openshift-origin-auth-remote-user/config/initializers/openshift-origin-auth-remote-user.rb
@@ -1,0 +1,3 @@
+Broker::Application.configure do
+  config.auth[:trusted_header] = "REMOTE_USER"
+end
