+++ swingshift/remote-user/openshift-origin-auth-remote-user.conf.sample
@@ -16,3 +16,16 @@ LoadModule authz_user_module modules/mod_authz_user.so
     Deny from all
     Satisfy any
 </Location>
+
+# The following APIs do not require auth:
+<Location /broker/rest/application_templates>
+    Allow from all
+</Location>
+
+<Location /broker/rest/cartridges>
+    Allow from all
+</Location>
+
+<Location /broker/rest/api>
+    Allow from all
+</Location>
